+++ telemetry.ui/CHANGELOG.md
@@ -4,8 +4,8 @@ The following tools were used to generate this project:
 - create-vite
 
 The following steps were used to generate this project:
-- Create react project with create-vite: `npm init --yes vite@latest telemetry.ui -- --template=react  --no-rolldown --no-immediate`.
-- Updating `vite.config.js` with port.
+- Create react project with create-vite: `npm init --yes vite@latest telemetry.ui -- --template=react-ts  --no-rolldown --no-immediate`.
+- Updating `vite.config.ts` with port.
 - Create project file (`telemetry.ui.esproj`).
 - Create `launch.json` to enable debugging.
 - Add project to solution.